+++ src/main/java/subway/controller/SectionController.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import subway.application.SectionService;
-import subway.ui.dto.request.SectionDeleteRequest;
-import subway.ui.dto.request.SectionUpdateRequest;
-import subway.ui.dto.response.SectionResponse;
-import subway.ui.dto.response.StationResponse;
+import subway.dto.request.SectionDeleteRequest;
+import subway.dto.request.SectionUpdateRequest;
+import subway.dto.response.SectionResponse;
+import subway.dto.response.StationResponse;
 
 @RestController
 @RequestMapping("/sections")
