+++ src/main/java/subway/dto/request/SectionDeleteRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.request;
+package subway.dto.request;
 
 public class SectionDeleteRequest {
 	private String upStationName;
