+++ src/main/java/subway/dto/request/SectionUpdateRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.request;
+package subway.dto.request;
 
 public class SectionUpdateRequest {
 	private String lineName;
