+++ src/main/java/subway/dto/response/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.response;
+package subway.dto.response;
 
 import subway.domain.Line;
