+++ src/main/java/subway/dto/response/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.response;
+package subway.dto.response;
 
 import subway.domain.Station;
